Add .cirun.yml to declare the OS configurations we want to use.
Stop using the Actions variables set in GitHub since they are single
values and not a list, with our self-hosted VM, we need a list of two
items at the very least, "self-hosted" and the VM configuration that
cirun manages.

It does result in more copy paste but it's trivial to grep, replace with
sed and just understand what is going on.